requirements.txt
gradio>=4.44.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
numpy>=1.24.0
pytesseract>=0.3.10
openai>=1.3.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0